+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 4.2.25
 + [x] Sidebar: remove scroll-y from sidebar, add scroll-y to sidebar-menu
++ [x] Countdown: refactoring structure, add animation effects: slide, fade, zoom
++ [x] Hero: minor upd styles for background image
 
 ### 4.2.24
 + [x] Time picker: fix method `val` for issue #1221