+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 + [x] Utils: add function `isLocalhost()`
 + [x] Docs: upd for using `htmlcontainer` component
 + [x] Window: fix gradually disappear for children when window is hiding, issue #1222
-+ [x] Utils: add methods `getCorsorPosition(...)`, `getCorsorPositionX(...)`, `getCorsorPositionY(...)` 
++ [x] Utils: add methods `getCursorPosition(...)`, `getCursorPositionX(...)`, `getCursorPositionY(...)` 
 + [x] Image compare: new component
 + [x] Image magnifier: new component